v0.90a - fixed syntax

+++ avimetrybot/cogs/counting.py
@@ -15,26 +15,27 @@ class Counting(commands.Cog):
         if message.guild is None:
             return
         countdoc = self.avimetry.collection.find_one({"_id": "counting"})
-        if str(message.guild.id) in countdoc:
-            if message.author.bot:
-                await message.delete()
-                print("bot")
-            elif message.author == self.avimetry.user:
-                print("self")
-                return
-            elif message.content != str(countdoc[str(message.guild.id)]):
-                print(countdoc[str(message.guild.id)])
-                await message.delete()
+        if message.channel.name == "counting":
+            if str(message.guild.id) in countdoc:
+                if message.author.bot:
+                    await message.delete()
+                    print("bot")
+                elif message.author == self.avimetry.user:
+                    print("self")
+                    return
+                elif message.content != str(countdoc[str(message.guild.id)]):
+                    print(countdoc[str(message.guild.id)])
+                    await message.delete()
+                else:
+                    guild=str(message.guild.id)
+                    newcount={guild: 1}
+
+                    self.avimetry.collection.update_one({"_id":"counting"}, {"$inc": newcount})
+
             else:
                 guild=str(message.guild.id)
-                newcount={guild: 1}
-
-                self.avimetry.collection.update_one({"_id":"counting"}, {"$inc": newcount})
-
-        else:
-            guild=str(message.guild.id)
-            newguild={guild: 0}
-            self.avimetry.collection.update_one({"_id":"counting"}, {"$set":newguild})
+                newguild={guild: 0}
+                self.avimetry.collection.update_one({"_id":"counting"}, {"$set":newguild})
 
     @commands.Cog.listener()
     async def on_message_edit(self, message_before, message_after):
